Add new class for responsibility of merging watershed fragments
